requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
twilio==9.8.7
urllib3==2.5.0
yarl==1.22.0
